nz-wp-newsletter

+++ wp-content/plugins/nz-wp-newsletter/wizard.php
@@ -46,7 +46,7 @@ class NzWpNewsletterWizard
         $sql = "CREATE TABLE $table_name (
               id mediumint(9) NOT NULL AUTO_INCREMENT,
               time datetime DEFAULT '0000-00-00 00:00:00' NOT NULL,
-              user_id mediumint(9),
+              user_id int,
               email varchar(55) NOT NULL,
               UNIQUE KEY email (email)
         ) $charset_collate;";
